+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Unreleased
 
 * Fixed:
+  * If two consecutive differential backups both ignored *different* save files
+    *and* none of those files were ignored in the associated full backup,
+    then the second differential backup would fail to redeclare
+    the first differential backup's ignored saves.
   * GUI: On Mac, the file/folder selector would cause the app to crash.
 
 ## v0.24.2 (2024-06-28)